requirements.txt
alpaca-py==0.21.1
annotated-types==0.6.0
backtrader==1.9.78.123
beautifulsoup4==4.12.3
certifi==2024.2.2
charset-normalizer==3.3.2
contourpy==1.2.1
cycler==0.12.1
fonttools==4.51.0
frozendict==2.4.4
html5lib==1.1
idna==3.7
kiwisolver==1.4.5
lxml==5.2.2
matplotlib==3.9.0
msgpack==1.0.8
multitasking==0.0.11
numpy==1.26.4
packaging==24.0
pandas==2.2.2
peewee==3.17.5
pillow==10.3.0
platformdirs==4.2.2
pydantic==2.7.1
pydantic_core==2.18.2
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
requests==2.31.0
six==1.16.0
soupsieve==2.5
sseclient-py==1.8.0
typing_extensions==4.11.0
tzdata==2024.1
urllib3==2.2.1
webencodings==0.5.1
websockets==12.0
yfinance==0.2.40
